Centralize visual identity logic and synchronize InfoScreen previews with in-game visuals
- **Visual Consistency**: Unified booster rendering by extracting `drawBoosterShape` into a shared internal function. The `InfoScreen` now uses `BoosterToBoosterUIMapper` to ensure that silhouettes, colors, and TopEnd glyph badges in the "Bách Khoa Vật phẩm" tab match in-game pickups 1:1.
- **Centralized Resolvers**: Introduced `BossKindResolver`, `ShipShapeColorMap`, and `BulletTypeColorMap` as single sources of truth. This eliminates hardcoded hex literals and redundant logic across `DialogShipPicker`, `GameState` banners, `StoryRegistry` taunts, and `InfoScreen` tabs.
- **Booster Identity Overhaul**: Refactored `BoosterToBoosterUIMapper` to use a `BoosterPreviewSpec` model. This ensures all 27 booster types have unique visual identities, specifically fixing color and shape collisions in the "ultimate," "health," and "laser" families.
- **Refactoring**:
    - Replaced redundant boss-resolution logic in `EnemyFactory`, `GameState`, and `StoryRegistry` with delegates to `BossKindResolver`.
    - Cleaned up manual coordinate math in `EnemyCanvas` and `InfoScreen`.
    - Improved `BoosterToBoosterUIMapper` internal constants to provide better contrast and disambiguation for non-base boosters.
- **Testing**: Added extensive unit tests to verify the visual contract:
    - `BoosterGlyphTest`: Ensures unique and valid Unicode glyphs for all boosters.
    - `BoosterPreviewColorDistinctTest`: Confirms all 27 booster types have distinct preview colors.
    - `BossKindResolverTest`: Validates chapter-aware boss overrides.
    - `BulletColorIdentityTest` & `ShipShapeColorMapTest`: Guards against color drift between UI components.
- **Tooling**: Added Gradle tasks to `.claude/settings.local.json` for targeted testing and compilation.

+++ app/src/main/java/com/tranphuloi/neon/ui/game/world/EnemyCanvas.kt
@@ -2131,7 +2131,7 @@ private fun DrawScope.drawEnemyTentacleSquid(cx: Float, cy: Float, w: Float, h: 
         val bx = cxLocal + (headR * kotlin.math.cos(a)).toFloat()
         val by = cyLocal + (headR * kotlin.math.sin(a)).toFloat()
         val ex = bx + (headR * 0.85f * kotlin.math.cos(a)).toFloat()
-        val ey = by + (headR * 1.20f).toFloat()
+        val ey = by + headR * 1.20f
         val tPath = Path().apply {
             moveTo(bx, by)
             cubicTo(bx + (i - 2.5).toFloat() * w * 0.05f, by + h * 0.10f,
